+++ database/migrations/2024_11_25_162638_create_product_purchase_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->foreignId('product_id')->constrained();
             $table->foreignId('purchase_id')->constrained();
             $table->primary(['product_id', 'purchase_id', 'unit_price']);
-            $table->decimal('quantity');
-            $table->decimal('unit_price');
+            $table->bigInteger('quantity');
+            $table->bigInteger('unit_price');
             $table->timestamps();
         });
     }
