+++ database/migrations/2024_11_25_165453_create_product_quotation_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->foreignId('quotation_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->primary(['product_id', 'quotation_id']);
-            $table->decimal('quantity');
-            $table->decimal('unit_price');
+            $table->bigInteger('quantity');
+            $table->bigInteger('unit_price');
             $table->timestamps();
         });
     }
