+++ database/migrations/2024_11_25_170001_create_order_product_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->primary(['product_id', 'order_id']);
-            $table->decimal('quantity');
-            $table->decimal('unit_price');
+            $table->bigInteger('quantity');
+            $table->bigInteger('unit_price');
             $table->timestamps();
         });
     }
