+++ database/migrations/2024_11_25_170021_create_credit_note_product_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->foreignId('credit_note_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->primary(['product_id', 'credit_note_id']);
-            $table->decimal('quantity');
-            $table->decimal('unit_price');
+            $table->bigInteger('quantity');
+            $table->bigInteger('unit_price');
             $table->timestamps();
         });
     }
