+++ database/migrations/2024_11_25_170037_create_invoice_product_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->foreignId('invoice_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->primary(['product_id', 'invoice_id']);
-            $table->decimal('quantity');
-            $table->decimal('unit_price');
+            $table->bigInteger('quantity');
+            $table->bigInteger('unit_price');
             $table->timestamps();
         });
     }
