+++ database/migrations/2024_12_13_021609_create_sale_sale_payment_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('sale_sale_payment', function (Blueprint $table) {
             $table->foreignId('sale_id')->constrained(); //50,000
             $table->foreignId('sales_payment_id')->constrained(); //40,000
-            $table->decimal('amount'); //20,000
+            $table->bigInteger('amount'); //20,000
             $table->timestamps();
         });
     }
